Adds WiremockNetworkTrafficListeners
Provides factory methods for creating arbitrary notifying traffic listeners.
